argus-insight-extensions/argus-insight-usersync/requirements.txt
PyYAML>=6.0
python-ldap>=3.4
sqlalchemy>=2.0
psycopg2-binary>=2.9
pymysql>=1.1
